kanche-share/app/resize.py
#!/usr/bin/python
# -*- coding: UTF-8 -*-

from PIL import Image
from StringIO import StringIO
from copy import copy
from img_rotate import fix_orientation

def resize(fin, box, fit, out):
    '''Downsample the image.
    @param img: file-like-object
    @param box: tuple(x, y) - the bounding box of the result image
    @param fix: boolean - crop the image to fill the box
    @param out: file-like-object - save the image into the output stream
    '''
    #preresize image with factor 2, 4, 8 and fast algorithm
    #img = Image.open(fin)
    data1 = fin.read()
    data2 = copy(data1)
    fin1 = StringIO(data1)
    fin2 = StringIO(data2)
    try:
        (img,_) = fix_orientation(fin1, save_over=False)
    except Exception as e:
        img = Image.open(fin2)
    factor = 1
    while img.size[0]/factor > 2*box[0] and img.size[1]*2/factor > 2*box[1]:
        factor *=2
    if factor > 1:
        img.thumbnail((img.size[0]/factor, img.size[1]/factor), Image.NEAREST)

    #calculate the cropping box and get the cropped part
    if fit:
        x1 = y1 = 0
        x2, y2 = img.size
        wRatio = 1.0 * x2/box[0]
        hRatio = 1.0 * y2/box[1]
        if hRatio > wRatio:
            y1 = int(y2/2-box[1]*wRatio/2)
            y2 = int(y2/2+box[1]*wRatio/2)
        else:
            x1 = int(x2/2-box[0]*hRatio/2)
            x2 = int(x2/2+box[0]*hRatio/2)
        img = img.crop((x1,y1,x2,y2))

    #Resize the image with best quality algorithm ANTI-ALIAS
    img.thumbnail(box, Image.ANTIALIAS)

    #save it into a file-like object
    img.save(out, "JPEG", quality=90)
#resize
